deleting test
Poorly placed tests set bad precedent also it wasnt really useful
